Param are of type float

+++ csc133-object-oriented-computer-graphics-programming/lab/lab01-SkyMail3000-part01/src/org/csc133/a1/GameObject.java
@@ -66,7 +66,7 @@ public abstract class GameObject {
      * @param newX
 	 * @param newY
      */
-    public void setLocation(double newX, double newY) {
+    public void setLocation(float newX, float newY) {
         this.location.set(0, newX);
 		this.location.set(1, newY);
     }
